+++ docs/changelog.md
@@ -5,6 +5,7 @@
 ### Dependencies
 
 - Bump [Pillow](https://github.com/python-pillow/Pillow) to v9.5.0
+- Bump [py7zr](https://github.com/miurahr/py7zr) to v0.20.4
 
 ## v2.8.3