+++ docs/changelog.md
@@ -7,6 +7,7 @@
 - Bump [Pillow](https://github.com/python-pillow/Pillow) to v9.5.0
 - Bump [py7zr](https://github.com/miurahr/py7zr) to v0.20.4
 - Bump [orjson](https://github.com/ijl/orjson) to v3.8.9
+- Bump [lxml](https://github.com/lxml/lxml) to v4.9.2
 
 ## v2.8.3